Changes BotCatMaxy/Startup to BotCatMaxy/Services and adds a TempActions folder to have TempActions.cs moved and renamed to TempActionService.cs
