+++ src/BotCatMaxy/Services/StatusManager.cs
